Red_Letter_Holidays/newrequirement.txt
certifi==2019.9.11
gunicorn
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.40.0
django-braces==1.13.0
django-cors-headers==3.1.1
django-historicalrecords==1.1
django-rest-passwordreset==1.0.0
djangorestframework==3.10.3
djangorestframework-custom-exceptions==1.0.1
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
pkg-resources==0.0.0
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.0
sqlparse==0.3.0
urllib3==1.25.6
